requirements.txt
# CyberSecurity Tools Hub - Python Dependencies
# Install with: pip install -r requirements.txt

# Network Tools
requests>=2.28.0
dnspython>=2.3.0
python-whois>=0.8.0
scapy>=2.5.0

# Web Tools
beautifulsoup4>=4.12.0
lxml>=4.9.0

# SSL/TLS
pyopenssl>=23.0.0
cryptography>=40.0.0

# QR Code Generator
qrcode>=7.4
pillow>=9.5.0

# Steganography
numpy>=1.24.0

# Optional - for enhanced functionality
user-agents>=2.2.0
colorama>=0.4.6

# Async HTTP for brute force tools
aiohttp>=3.8.0
certifi>=2023.0.0

# Screenshot capture (optional)
# selenium>=4.0.0
# playwright>=1.30.0
